feat: padronizando descrição dos testes de acordo com a iJunior
